requirements.txt
Django==3.1
django-filter
djangorestframework==3.12.4
Markdown
django-cors-headers
djangorestframework-camel-case